+++ lib/roda/plugins/data_grid.rb
@@ -157,6 +157,7 @@ class Roda
         def search_presenter(id, dmc, params, for_rerun)
           OpenStruct.new(rpt: dmc.report,
                          qps: dmc.parameter_list,
+                         caption: dmc.render_options[:caption],
                          rpt_id: id,
                          load_params: (params[:back] && params[:back] == 'y'),
                          rerun: for_rerun)